Use size_t for data length variables.

+++ src/demuxers/demux_mod.c
@@ -73,7 +73,7 @@ typedef struct {
   char                *title;
   char                *artist;
   char                *copyright;
-  off_t                filesize;
+  size_t               filesize;
   
   char                *buffer;
